change: renamed shap_wrappers.py to solve merge conflict with master
